pyproject.toml
[project]
name = "course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=4.2.27",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "djangorestframework-simplejwt>=5.3.1",
    "django-cors-headers>=4.4.0",
    "psycopg2-binary>=2.9.9",
    "psycopg2>=2.9.11",
]
